add: unique design style, photos, correct components
