+++ src/Footer.js
@@ -3,7 +3,7 @@ import './Footer.css'
 
 export const Footer = () => {
     return (
-        <div className="item_3 item_md_8 item_lg_12">
+        <div className="item_3 item_md_8 item_lg_12 mt6">
                     <div className="card footer">
                         <div className="grid">
                             <div className="item">
