+++ src/Header.css
@@ -2,6 +2,13 @@
     font-family: 'Bitter', serif;
 }
 
+h2.showOn {
+    color: rgb(255 255 255 / 94%);
+    font-size: 3rem;
+    text-shadow: 1px 4px 6px #b7b7b7, 0 0 0 #000, 1px 4px 6px #def;
+    user-select: none;
+}
+
 .boxShadow {
     box-shadow: 1px 1px 15px #eaeaea, -1px -1px 15px #eaeaea;
     margin: 10px;
@@ -10,19 +17,59 @@
 }
 
 .navLink {
-    font-weight: 500;
-    font-size: 18px;
+    font-weight: 600;
+    font-size: 20px;
     text-transform: capitalize;
+    position: relative;
+    letter-spacing: 1px;
+    color: #fff;
+    padding: 2px 10px;
+}
+.navLink::before {
+    content: 'home';
+    position: absolute;
+    left: -3px;
+    top: -2px;
+    color: var(--greeny);
+    transition: all 0.3s ease;
+}
+.navLink:nth-child(2)::before{
+    content: 'sources';
+}
+.navLink:nth-child(3)::before{
+    content: 'Video Gallery';
+}
+.navLink:nth-child(4)::before{
+    content: 'Bookshelf';
+}
+.navLink:nth-child(5)::before{
+    content: 'About Project';
+}
+.navLink:hover::before {
+    color: rgba(84, 84, 84, 0.2);
 }
 
+
 .imageLogo_container {
-    width: 390px;
+    max-width: 390px;
+    width: 100%;
     height: 200px;
     border-radius: 5px;
     overflow: hidden;
-    filter: drop-shadow(2px 4px 6px #4e4e4e);
+    margin: 0 auto;
+    opacity: 0.8;
+    position: relative;
 }
 
 .imageLogo{
-
+}
+.imageLogo_container::after {
+    position: absolute;
+    content: '';
+    top: -2rem;
+    left: -2rem;
+    width: 30%;
+    height: 120%;
+    background-color: #f4f4f4;
+    transform: rotate(20deg);
 }
