+++ src/pages/index.js
@@ -1,2 +1,3 @@
 export * from './About'
+export * from './Home'
 export * from './Sources'
